Interrupt as constructor. Context for causes.
Update so that `Interrupt` is a constructor. Change the function
signature so that you must specify the `cause` or `causes` in the
`context` object. They will be extracted and printed nested in a
heirachrical layout with their stack traces.

Add the ability to specify context for causes. This allows you to print
contextual information known by the catcher in addition to the
contextual information added to the thrown exception. It makes
invocation with an array of causes fussier, but if you're doing an array
of causes you're already fussing because you're maintaining an array of
caught exceptions. The causes will now be specified with an array of
arrays, instead of an array of causes. The arrays in the array will have
one or two elements. The first element is the cause. The second,
optional element is a context object that will be rendered as JSON5.

There is no more shorthand `cause` property. It is always an array of
`causes` and a parallel array of `contexts`.