Make additional properties non-iterable.
Suddenly in Node.js 12 the default exception handler is printing
iterable properties so that our nicely formatted nested stack trace is
followed by a dump of jumble of nested JSON-esque data. Able to prevent
this by assigning the properties add to the `Interrupt` instance using
`Object.assignProperty()` with the `iterable` flag set to `false`.

Closes #78.