Implement prune of repeated causes.
Implement a dedup of repeated causes. If a cause is repeated more than
once it the first cause is used a `repeated` property with a count of
instances is added to the cause context.

Will prune entire trees if they have the same structure and causes.

This can potentially lose some specific data, say removing a repeated
failure to open files originating from the same source file and line,
but opening different file name.

The default determination for the uniqueness of an error is the source
file and line number, but the user can provide a different function to
pluck a key out of the errors in the tree.

Closes #101.