Promise resolver, callback helper function.
Getting really particular about having the correct line number in a
stack trace. It is so disheartening to see those stack trace free errors
on the command line and worse, in the logs. The solutions are not pretty
and this module is becoming ever more specific to my particular brand of
madness. Going to have to try to write the documentation as advocacy to
those who truly want to listen, because it unless you're able to
establish a fashion, it all appears to be excessive.

Created a resolve function that will wrap a Promise in a try/catch block
and if it raises an exception it will wrap the error in an `Interrupt`
derrived error. Found that this would not actually report the call stack
from which the `async` function was invoked. Not unless you're using
Node.js 14 or greater.

Created a callback function builder that builds a callback that will
wrap an error in an `Interrupt` derived error. It uses a function that
calls a function to construct the exception so that the file and line of
the application source appears in the stack trace.