Create `Symbol`s for error codes.
Now creating `Symbol`s for error codes. For each code an property with
the name of the code and the generated `Symbol` is set on the generated
`Error` class. When an exception is constructed it has a non-enumerable
`symbol` property set. The `symbol` will not appear in the stack trace,
it is already represented by the `code` string. You can specify the code
as the first argument to the exception constructor with the code name or
the `Symbol`.

Closes #124.