Sketch auditing, deferred construct. Assert docs.
Documenting `assert` for the sake of getting the deferred construction
right. There is an ambiguity with deferred construct in that it is
always just a function but callee is also a function. I'm creating many
special cases in the code trying to not interpret it as a callee.

I want to provided memoized resolvers, callbackers, etc. While
implementing one I made the `callee` an option and realized that it is a
rather uncommon option and could therefore always be an option. This
would be easier to document. There is no way to specify a callee as a
positional parameter, you must use an option object.

Did get an initial `audit` wrapper. Decided that it is enough to just
call an audit function since this is only ever intended for use in unit
testing, ever for use in production, never to test the exceptions of a
dependency. If those use cases arise you can use the audit facility, but
it should not be as simple as an environment variable, as if it was a
dormant production facility that ready to be switched on and off.

Going to update the documentation first, to see if it makes life easier,
then continue with the implementation of the wrapper functions.

See #117.
See #114.
See #97.