Settle on an argument order for constructor.
Seem to have removed ambiguities. The first argument to the constructor
must be an options array, code symbol, code string, or message format.
After that arguments can appear in any order. The options object versus
properties object ambiguity is resolved by the position. The properties
object cannot appear as the first argument and the options object can no
longer be repeated for merging.

Still need to be able to merge options objects programmatically, though.
Will create a separate function for this.

See #117.