Sketch of currying, implement `voptions`.
Added the ability to curry `resolve`. Will apply it to all helper
functions that raise exceptions except assert. Well, maybe assert still,
but it is going to be the place where the syntax bashing falls apart.

Ambiguities appear to be resolved. For currying we tell the user to call
with an appropriate first argument, options object, code string, code
symbol or message format, and to omit what the argument is supposed to
be which is either a `Promise` or a function. We can be sure that we're
not mistaking an options object for a `Promise` unless the user decides
to add a `then` function property for no good reason. We're not going to
document an admonisment not to do so, it won't happen.

Closes #126.
See #111.
See #125.