Rename `context` to `properties`.
The separation was established when I sought to allow the user to choose
which properties appear in the stack. Now they will all appear. If the
user wants to add additional properties that do not appear in the stack
trace they can assign them. It's nothing I never need to do.

Closes #129.