Fix options merging.
Was prepsending the previous options object to variable argment array to
use the varabile argument merge, but if the first argument to the
variable argument was already an object it would be interpreted as
properties.

Rewrote the options merge/interpreter. Let the old one in place because
I have something `git stash`ed that I need to merge.

See #117.