Sketch of `AggregateError` support.
Arriving in Node.js 15, but the current shim doesn't work correctly for
reasons I don't care to investigate. This was a nice to have and it
doesn't really matter if an `Interrupt` is an `AggregateError`. I'd have
to be convinced from someone who adopted Interrupt, not as an advocacy
effort.