Sort out enumerability.
Sorted out enumerability, which is whay I meant by un-overridable. Well,
also sorted out how to prevent overrides of properties.

Updated `Interrupt.JSON.stringify()` to not include the message if the
root object is an `Interrupt` so that it displays the context message as
it appears in in the stack trace.

Additionally, added a section in the `readme.t.js` to look ahead at
documentation and get some test coverage. Looking at what needs to be
done to allow the user to create their own assertions and wrappers.
Using coverage to see what corner cases need to be documented.

Closes #128.