Re-implementation of stack trace parser.
Going to ensure that the parser works and also add a scanning ability to
the parser. Start of parser is here. In order to scan the parser
instance will accept a line at a time and return an object when it's
`end()` method is called. Parsing nested exceptions is a matter of
constructing a nested parser instance.

Requires reworking of how we serialize non-Error errors which will
probably be just JSON. The only special case is `null` since someone
could create a stack-traceless error with a name of `null` which would
be really unpleasant, highly-unlikely but if it is the only special case
we should accommodate it.

Alternatively, we could detect that the `error.stack` evaulates to
`null` and print `stack:` or a properties hash of `'{\n}'`.

See #102.