Inheriting templates from super class.
Don't imagine that people are going to create error classes from other
modules, but if they do, they should be careful not to use existing
templates since they are not public. Seems like maybe they should not be
inherited, though. The user can always define a set of common templates
in a module somewhere, as a builder function and import it to different
modules in their code, so I'm liable to back out of this inheritance.

Come to think of it, the messages are not public either. If you where to
inherit the messages associated with codes you'd end up having to
document the message formats. It really does make more sense to have a
builder function and to specify your own symbols.

Currently you could use someone else's `Error` class, like `TypeError`,
but you can't use their formatting.

See #140.