`create` unshifts onto `vargs`.
Working through code definition and started with updating `create`'s
`vargs` processing so that an array can include anything that is an
argument and gets unshifted onto the `vargs` so that definitions are
recursive. This means code functions can return arrays.

Also, there's an ambiguity between an initial code function and the
super class. They are both functions.

See #140.
See #133.