Options object can now go anywhere.
No distinction between the options object and the properties objects.
Special arguments like `callee`, `type` and `_errors` are now
`'#callee'`, `'#type'`, and `'#errors'` whose naming convention
corresponds to the new private member feature of JavaScript. Easier to
document than the distinction between the two type-less objects, that
distinction being only their position the positional arguments.

Re-implemented options "parsing" but with preservation of property
traits (such as `enumerable`) and copious error checking and reporting.
The actual structure of an error is to be determined, but the conditions
are all there so when the time comes it is search and replace.

Now it ought to be easier to sort how to implement all the code
inheritance nonsense that I'll never use but ought to be there so that I
could use it if I wanted to. That is, so that its usefulness is
discoverable.

See #117.