Inheritance and aliases.
It's almost sorted out. Found an inheritance mechanism I'm willing to
field questions about. We'll talk about how it works in the
documentation.

Internally, prototypes are flattened now. They are not constructed using
`options` but `combine` which does a simpler merge of object preserving
their property descriptors. Greatly simplifies the code to not have to
assemble an object from `code`, `symbol` and `properties`, to have them
flattened together. We've made certain that `code` and `symbol` are
correct in the prototype. Aliases should have no `symbol` property in
the prototype, they should inherit from the aliased prototype.

Aliases are assembled at construction. We don't pre-merge them. This
will allow someone to design a set of aliases with a middle alias that
can be overridden in a descendent.

Straightened out the `Prototype` object. Determined a set of necessary
properties for said object. Came up with a set of names that I found
suitable. Adjusted it for the new flattened prototypes.

Fixed a regular expression. It was of the form `/^a|b$/` and had to be
rewritten as `/^(?:a|b)$/`. I recall the former being a valid Perl
regular expression, but it isn't working in JavaScript. Tried searching
for details on Google no matter how I specific I got, I couldn't get
anything but page upon page of JavaScript regular expression tutorials
and I'm not up for...

Oh, it just occurred to me. It's valid but its wrong. It matches `^a` OR
`b^` so it will match `'cb'`. The non-capturing group is therefore
required. Obvious, really.