Implement `toString()` and `inspect.util`.
`inspect.util` is now dumping enumerable `Error` properties at the end
after the stack trace. It allows for the specification of a special
function that `inspect.util` will use to get a value to dump so we use
it to return the stack.

Implemented a `toString()` that returns the name and message if it
exists and looks like `Error.toString()`.

Closes #142.