User usable `poke`.
Add the ability for the user to create their own poker interfaces. A
useful trace into the user code may not be at the point where `assert`
or `resolve` or the like are called.

This implementation requires that the user create their own variant of
`assert`, `resolve`. How would we create one where the poker was an
option in the options object of the constructor?