Start of a universal poker format.
A poker becomes `$ => $()` and that is all it ever is. It does not
provide arguments. Now we can simply check that the poker was called,
not whether it correctly provided arguments, correctly returned the
results of constructor function. It must always be the last argument to
construct. Users can pass in a poker from the edge of their interface.

Before proceeding with this, I want to implement a `#vargs` property on
the options object so that `Class.options` can be called without having
to call `Array.concat` if one of the arguments is a variadic argument
array.