Thoughts on making nested pokers.
Looking for a way to have a stack of pokers. Think I found it. It would
be a point in the stack trace where the stack reverses. The top of the
stack trace is the place where the pokers are called, then above that
you go back up the stack to places where pokers where provided. This
would allow you to trace back up into application code to where an
asynchronous operation was initiated.

This is still a dubious construct, but worth exploring. I have to catch
a bug with with it at some point to be enthusastic about. Seems
promissing, though.