Replace `#` prefix with `$`.
Makes the code nicer. Makes it easier for users to use special features,
they don't have to quote. Feels like defensive programming to use `'#'`
as if someone is going to mistakenly use `'$stack'` as a special
variable. Perhaps we allow all `$` prefixes and intercept the ones we
document as special.