Fix callee now that deferred construction is gone.
We removed deferred construction so we no longer treat the poker
function as a deferred construction function in the case of `assert` and
`invoke` which means we no longer have to tell construct to strip the
stack trace down to `assert` or `invoke` instead of stripping it to the
deferred constructor.

With this we've created a static `create()` function we could actually
use in our applications.