Tracer resolution in constructor.
It's already in the constructor but now, quite obviously, we can move
the invocation of the tracer functions into the constructor.

Now you can add tracers from any error constructor so we've added them
to the variadic arguments. We've also added deferred constructor
functions. The distinction is in the function signature. No arguments is
a deferred constructor and one argument is a tracer.

It also greatly localizes the use of the Google V8 specific features.
They are now all in the constructor. It would be easier to accommodate a
port should it arise.

Which makes me want to comment on the GitHub Issues in my head. Look how
you get to a point of accommodation regardless of whether you set out to
address the GitHub Issues in your head, or dismiss them. It ought to be
considered a pre-mature optimization to consider webpack.