Stringify AssertionError.
AssertionError adds the code to the name, freezes the stack, then sets
the name again. Made a note of this in `swipe.t.js` with the test; that
it would be good to mention as part of advocacy, that Node.js is itself
as tricky as Interrupt and as dependent on the Google V8 Error
implementation.