Not ready to delete old `readme.t.js`.
The old `readme.t.js` is a lot of work, a month and a half, and it was
written for its format, so I don't know that you want to just copy and
paste it into Edify and expect it to be more compelling because it is a
`README.md`. You probably want a succinct readme and then rename
`readme.t.js` to `tutorial.t.js` or something.